+++ src/main/kotlin/com/_week/hhplus_tdd_practice/infra/LectureRepository.kt
@@ -1,6 +1,6 @@
 package com._week.hhplus_tdd_practice.infra
 
-import com._week.hhplus_tdd_practice.domain.entity.Lecture
+import com._week.hhplus_tdd_practice.infra.entity.Lecture
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
